fix: address critical findings from review (spike_sorting + batch_jobs)
Spike sorting (8 critical fixes):
- sorting_extractor: move pandas import into __init__ with actionable
  ImportError; add opt-in compact=True to remap sparse Phy cluster_ids
  to dense 0..N-1, preserving original_unit_ids for traceability and
  fancy-indexing templates.npy via get_templates_all.
- ks2_runner: drop dead natsort/scipy.io imports; gate spikeinterface
  imports behind try/except inside the functions that use them
  (BaseRecording under TYPE_CHECKING with quoted annotations); fix
  Windows AttributeError by using getattr(signal, "SIGKILL",
  signal.SIGTERM) in stop() and Popen.kill() in kill().
- waveform_extractor: replace exec() with ast.parse + ast.literal_eval
  for params.py (closes RCE — only literal assignments and
  range(literal_ints) accepted); wire _check_unit_id_density guard
  into the two template-cache allocations.
- sorting_utils: new _check_unit_id_density helper raises a clear
  MemoryError when max(unit_ids) > 100 * len(unit_ids), telling the
  user the GB cost and how to opt into compaction; add sorter_name
  param to get_paths so each backend writes to {sorter_name}_results
  instead of all sharing kilosort2_results (silent cross-sorter cache
  corruption).
- pipeline: _get_noise_levels falls back to a single full-trace MAD
  estimate with UserWarning when length <= chunk_size; raises
  ValueError on empty input; passes sorter_name to get_paths.
- maxwell_io: astype(dtype, copy=True) prevents in-place mutation of
  caller buffer when dtype == raw_u16.dtype; peak RAM unchanged in
  the typical float32 path.
- backends/rt_sort: wire _check_unit_id_density guard ahead of the
  synthetic templates allocation.

Batch jobs (3 critical fixes):
- cli: _apply_image_selection raises SystemExit with profile name,
  requested image_profile, and available alternatives instead of
  silently passing through to a generic Pydantic error.
- entrypoints/sorting: strict length check + zip iteration replaces
  min(...) saturation, so a sort_recording result-count mismatch
  fails loudly instead of silently clobbering or dropping pickles.
- session: new _s3_env_overlay() plumbs profile.endpoint_url and
  profile.region_name into the pod env at both workspace and sorting
  submission sites so NRP / non-AWS S3 clusters work end-to-end.

Minor (2 follow-ups from /complete_review):
- spikeslicestack: suppress RuntimeWarning ("Mean of empty slice")
  in responsive_units when units have no baseline spikes.
- data_loaders: replace start_time_ms or 0.0 with explicit
  ``if start_time_ms is not None`` in both NWB load paths.

+++ src/spikelab/spike_sorting/maxwell_io.py
@@ -248,7 +248,13 @@ def load_maxwell_native(
 
     # Scale uint16 -> microvolts.  The MaxWell convention is
     # ``volts = raw * lsb``, so µV = raw * (lsb * 1e6).
-    traces_uv = raw_u16.astype(dtype, copy=False)
+    # ``copy=True`` is required: with ``copy=False`` and a caller that
+    # passed ``dtype`` equal to ``raw_u16.dtype`` (e.g. ``"uint16"``),
+    # ``astype`` would return ``raw_u16`` itself and the in-place
+    # ``*=`` below would mutate the caller's buffer. In the common
+    # ``dtype="float32"`` path the astype is already a copy anyway, so
+    # peak RAM is unchanged.
+    traces_uv = raw_u16.astype(dtype, copy=True)
     traces_uv *= np.asarray(lsb_volts * 1e6, dtype=dtype)
     # (channels, samples) -> (samples, channels) for SpikeInterface
     traces_si = np.ascontiguousarray(traces_uv.T)
